moved DOM tests to separate file, deleted comments
